requirements.txt
discord.py>=2.3.2
jishaku>=2.5.0
aiohttp>=3.9.0
pytz>=2023.3
python-dotenv>=1.0.0
amari.py>=1.0.0
psutil>=5.9.6